feat: Resolve issues #4 and  #7
