+++ Cargo.toml
@@ -14,3 +14,6 @@ readme = "README.md"
 [dependencies]
 image = "0.25.5"
 walkdir = "2.5.0"
+
+[dev-dependencies]
+dotenv = "0.15.0"
