+++ src/frame.rs
@@ -1,0 +1,34 @@
+#[derive(Debug)]
+pub struct Frame {
+    x: u32,
+    y: u32,
+    width: u32,
+    height: u32,
+}
+
+impl Frame {
+    pub fn new(x: u32, y: u32, width: u32, height: u32) -> Self {
+        Self {
+            x,
+            y,
+            width,
+            height,
+        }
+    }
+
+    pub fn get_x(&self) -> u32 {
+        self.x
+    }
+
+    pub fn get_y(&self) -> u32 {
+        self.y
+    }
+
+    pub fn get_width(&self) -> u32 {
+        self.width
+    }
+
+    pub fn get_height(&self) -> u32 {
+        self.height
+    }
+}
